完善后台的管理员修改密码。重置商户密码，对商户账户的增删改查等

+++ app/Http/Controllers/UserController.php
@@ -4,21 +4,28 @@ namespace App\Http\Controllers;
 
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
 
 class UserController extends Controller
 {
     //
     public function index()
     {
-        $users=User::all();
+        $users=User::paginate(6);
 
         return view('user.index',compact('users'));
     }
 
-    public function create()
-    {
 
+    public function updatestatus(User $user,Request $request)
+    {
+        $user->update(['status'=>$request->status]);
+
+        session()->flash('success','操作成功');
+
+        return redirect()->route('users.index');
     }
+
 
     public function edit(User $user)
     {
@@ -29,12 +36,13 @@ class UserController extends Controller
     {
        $this->validate($request,
            [
-               'name'=>'required|max:15',
-               'email'=>'required|email',
+               'name'=>['required','max:15',Rule::unique('users')->ignore($user->id)],
+               'email'=>['required',Rule::unique('users')->ignore($user->id)],
                'password' => ['required', 'between:6,18', 'confirmed'],
            ],
            [
                'name.required'=>'账号不能为空',
+               'name.unique'=>'账号已存在',
                'name.max'=>'账号不能超过15',
                'email.required'=>'邮箱不能为空',
                'password.required'=>'密码必填',
@@ -56,7 +64,49 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        $success=['success'=>true];
+        if($user->shop_id){
+
+            $success=['success'=>false];
+            $res=json_encode($success);
+            echo $res;
+            return ;
+
+        };
+
         $user->delete();
-        echo '删除成功';
+        $res=json_encode($success);
+        echo $res;
     }
+
+    //重置密码页面
+
+    public function resetpass()
+    {
+        return view('user.repass');
+    }
+
+    public function resetname(Request $request)
+    {
+        $userinfo=User::where('name',$request->name)->first();
+        if(!$userinfo){
+            session()->flash('danger','账号不存在');
+            return redirect()->route('resetpass');
+        }
+
+        return view('user.repassword',compact('userinfo'));
+
+    }
+
+    public function resetpassword(User $user,Request $request)
+    {
+        $password=bcrypt($request->password);
+
+         $user->update(['password'=>$password]);
+
+        session()->flash('success','重置成功');
+
+        return redirect()->route('users.index');
+    }
+
 }
